Moved export_shapefile from raster to separate 'shapefiles' module.
